Fundoo

+++ application/service/LabelService.php
@@ -75,30 +75,72 @@ class LabelService extends CI_Controller
         return $arr;
     }
 
-    public function deletelname($id)
+    public function deletelname($id,$flag,$labelValue)
     {
-        $query = "DELETE FROM label WHERE id = '$id'";
+        if($flag == 'delete')
+        {
+            $query = "DELETE FROM label WHERE id = '$id'";
+            $stmt = $this->db->conn_id->prepare($query);
+            $res = $stmt->execute();
+            if ($res) {
+                $data = array(
+                    "status" => "200",
+                );
+                print json_encode($data);
+            } else {
+                $data = array(
+                    "status" => "204",
+                );
+                print json_encode($data);
+                return "204";
+            }
+        }
+        else if($flag == 'update')
+        {
+            if($labelValue != '' && $labelValue != 'undefined')
+            $query = "UPDATE label set label = '$labelValue' WHERE id = '$id'";
+            $stmt = $this->db->conn_id->prepare($query);
+            $res = $stmt->execute();
+            if ($res) {
+                $data = array(
+                    "status" => "200",
+                );
+                print json_encode($data);
+            } else {
+                $data = array(
+                    "status" => "204",
+                );
+                print json_encode($data);
+                return "204";
+            } 
+        }
+    }
+
+
+    public function labelNoteDis($id)
+    {
+        $query = "SELECT n.id,n.title,n.noteContent,n.date,n.color,n.image,n.pin,l.label from addnote n JOIN label_note ln ON ln.note_id=n.id JOIN label l on ln.label_id=l.id where l.id=$id";
         $stmt = $this->db->conn_id->prepare($query);
         $res = $stmt->execute();
-        if ($res) {
-            $data = array(
-                "status" => "200",
-            );
-            print json_encode($data);
-        } else {
-            $data = array(
-                "status" => "204",
-            );
-            print json_encode($data);
-            return "204";
+        $arr = $stmt->fetchAll(PDO::FETCH_ASSOC);
+        foreach($arr as $notes)
+        {
+            $title = $notes['title'];
+            $noteContent = $notes['noteContent'];
+            $date = $notes['date'];
+            $color = $notes['color'];
+            $image = $notes['image'];
+            $label = $notes['label'];
         }
+
+        print json_encode($arr);
     }
 
     public function addLtoN($note_id,$label_id)
     {
         $query = "INSERT into label_note (note_id,label_id) values ('$note_id','$label_id')";
 
-        $stmt = $this->db->conn_id->prepare($query);
+           $stmt = $this->db->conn_id->prepare($query);
             $res = $stmt->execute($data);
             // return $res;
 
@@ -119,25 +161,6 @@ class LabelService extends CI_Controller
                 return "204";
 
             }
-            return $result;
     }
-
-    public function labelNoteDis($id)
-    {
-        $query = "SELECT n.id,n.title,n.noteContent,n.date,n.color,n.image,n.pin,l.label from addnote n JOIN label_note ln ON ln.note_id=n.id JOIN label l on ln.label_id=l.id where l.id=$id";
-        $stmt = $this->db->conn_id->prepare($query);
-        $res = $stmt->execute();
-        $arr = $stmt->fetchAll(PDO::FETCH_ASSOC);
-        foreach($arr as $notes)
-        {
-            $title = $notes['title'];
-            $noteContent = $notes['noteContent'];
-            $date = $notes['date'];
-            $color = $notes['color'];
-            $image = $notes['image'];
-            $label = $notes['label'];
-        }
-
-        print json_encode($arr);
-    }
+    
 }
